+++ crates/wdl-ast/src/v1/expr.rs
@@ -2075,6 +2075,50 @@ pub enum LiteralStringKind {
     Multiline,
 }
 
+/// Represents a multi-line string that's been stripped of leading whitespace
+/// and it's line continuations parsed. Placeholders are not changed and are
+/// copied as-is.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum StrippedStringPart {
+    /// A textual part of the string.
+    Text(String),
+    /// A placeholder encountered in the string.
+    Placeholder(Placeholder),
+}
+
+/// Removes line continuations from a string.
+///
+/// A line continuation is a backslash immediately preceding a newline
+/// character. This function will remove the backslash and newline, and any
+/// spaces or tabs that follow the newline character.
+fn remove_line_continuations(s: &str) -> String {
+    let mut result = String::new();
+    let mut chars = s.chars().peekable();
+    let mut push_c;
+    while let Some(c) = chars.next() {
+        push_c = true;
+        if c == '\\' {
+            if let Some(next) = chars.next() {
+                if next == '\n' {
+                    push_c = false;
+                    let inner_chars = chars.by_ref();
+                    while let Some(&c) = inner_chars.peek() {
+                        if c == ' ' || c == '\t' {
+                            inner_chars.next();
+                        } else {
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+        if push_c {
+            result.push(c);
+        }
+    }
+    result
+}
+
 /// Represents a literal string.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct LiteralString(pub(super) SyntaxNode);
@@ -2124,6 +2168,116 @@ impl LiteralString {
         }
 
         None
+    }
+
+    /// Strips leading whitespace from a multi-line string.
+    ///
+    /// This function will remove leading whitespace from each line of a
+    /// multi-line string and parse line continuations. Single or double
+    /// quoted strings will return `None`.
+    pub fn strip_whitespace(&self) -> Option<Vec<StrippedStringPart>> {
+        if self.kind() != LiteralStringKind::Multiline {
+            return None;
+        }
+
+        let mut result = Vec::new();
+
+        // Parse the string parts and remove line continuations.
+        // We also remove the first and last lines of the string.
+        // Placeholders are copied as-is.
+        for (i, part) in self.parts().enumerate() {
+            match part {
+                StringPart::Text(text) => {
+                    let parsed = remove_line_continuations(text.as_str());
+                    let mut reconstructed = Vec::new();
+                    for (j, line) in parsed.lines().enumerate() {
+                        if i == 0 && j == 0 {
+                            let trimmed = line.trim_start();
+                            if !trimmed.is_empty() {
+                                reconstructed.push(trimmed.to_string());
+                            }
+                            continue;
+                        }
+                        reconstructed.push(line.to_string());
+                    }
+
+                    result.push(StrippedStringPart::Text(reconstructed.join("\n")));
+                }
+                StringPart::Placeholder(placeholder) => {
+                    result.push(StrippedStringPart::Placeholder(placeholder));
+                }
+            }
+        }
+
+        if let Some(StrippedStringPart::Text(text)) = result.last_mut() {
+            if let Some(index) = text.rfind(|c| !matches!(c, ' ' | '\t')) {
+                text.truncate(index + 1);
+            } else {
+                text.clear();
+            }
+            if text.ends_with('\n') {
+                text.pop();
+            }
+        }
+
+        // Now that the string has had line continuations parsed and the first and last
+        // lines removed, we can detect any leading whitespace and trim it.
+        let mut leading_whitespace = usize::MAX;
+        let mut parsing_leading_whitespace = true;
+        for part in &result {
+            match part {
+                StrippedStringPart::Text(text) => {
+                    for (i, line) in text.lines().enumerate() {
+                        if i > 0 {
+                            parsing_leading_whitespace = true;
+                        }
+                        if parsing_leading_whitespace {
+                            let mut ws_count = 0;
+                            for c in line.chars() {
+                                if c == ' ' || c == '\t' {
+                                    ws_count += 1;
+                                } else {
+                                    break;
+                                }
+                            }
+                            leading_whitespace = leading_whitespace.min(ws_count);
+                        }
+                    }
+                }
+                StrippedStringPart::Placeholder(_) => {
+                    parsing_leading_whitespace = false;
+                }
+            }
+        }
+
+        let mut parsing_leading_whitespace = true;
+        for part in &mut result {
+            match part {
+                StrippedStringPart::Text(text) => {
+                    let mut lines = Vec::new();
+                    for (i, line) in text.lines().enumerate() {
+                        if i > 0 {
+                            parsing_leading_whitespace = true;
+                        }
+                        if !parsing_leading_whitespace {
+                            lines.push(line);
+                            continue;
+                        }
+                        if line.len() < leading_whitespace {
+                            lines.push("");
+                        } else {
+                            lines.push(&line[leading_whitespace..]);
+                        }
+                    }
+                    *text = lines.join("\n");
+                }
+                StrippedStringPart::Placeholder(_) => {
+                    parsing_leading_whitespace = false;
+                }
+            }
+        }
+
+        Some(result)
     }
 }
 
@@ -7361,5 +7515,222 @@ task test {
         let mut visitor = MyVisitor(0);
         document.visit(&mut (), &mut visitor);
         assert_eq!(visitor.0, 1);
+    }
+
+    #[test]
+    fn strip_whitespace_on_single_line_string() {
+        let (document, diagnostics) = Document::parse(
+            r#"
+version 1.1
+
+task test {
+    String a = "  foo  "
+}"#,
+        );
+
+        assert!(diagnostics.is_empty());
+        let ast = document.ast();
+        let ast = ast.as_v1().expect("should be a V1 AST");
+
+        let tasks: Vec<_> = ast.tasks().collect();
+        assert_eq!(tasks.len(), 1);
+
+        let decls: Vec<_> = tasks[0].declarations().collect();
+        assert_eq!(decls.len(), 1);
+
+        let expr = decls[0].expr().unwrap_literal().unwrap_string();
+        assert_eq!(expr.text().unwrap().as_str(), "  foo  ");
+
+        let stripped = expr.strip_whitespace();
+        assert!(stripped.is_none());
+    }
+
+    #[test]
+    fn strip_whitespace_on_multi_line_string_no_interpolation() {
+        let (document, diagnostics) = Document::parse(
+            r#"
+version 1.2
+
+task test {
+    # all of these strings evaluate to "hello  world"
+    String hw1 = <<<hello  world>>>
+    String hw2 = <<<   hello  world   >>>
+    String hw3 = <<<   
+        hello  world>>>
+    String hw4 = <<<   
+        hello  world
+        >>>
+    String hw5 = <<<   
+        hello  world
+    >>>
+    # The line continuation causes the newline and all whitespace preceding 'world' to be 
+    # removed - to put two spaces between 'hello' and world' we need to put them before 
+    # the line continuation.
+    String hw6 = <<<
+        hello  \
+            world
+    >>>
+}"#,
+        );
+
+        assert!(diagnostics.is_empty());
+        let ast = document.ast();
+        let ast = ast.as_v1().expect("should be a V1 AST");
+
+        let tasks: Vec<_> = ast.tasks().collect();
+        assert_eq!(tasks.len(), 1);
+
+        let decls: Vec<_> = tasks[0].declarations().collect();
+        assert_eq!(decls.len(), 6);
+
+        let expr = decls[0].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[1].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[2].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[3].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[4].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[5].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  world"),
+            _ => panic!("expected text part"),
+        }
+    }
+
+    #[test]
+    fn strip_whitespace_on_multi_line_string_with_interpolation() {
+        let (document, diagnostics) = Document::parse(
+            r#"
+version 1.2
+
+task test {
+    String hw1 = <<<
+        hello  ${"world"}
+    >>>
+    String hw2 = <<<
+        hello  ${
+            "world"
+        }
+        my name
+        is \
+            Jerry\
+    !
+    >>>
+}"#,
+        );
+
+        assert!(diagnostics.is_empty());
+        let ast = document.ast();
+        let ast = ast.as_v1().expect("should be a V1 AST");
+
+        let tasks: Vec<_> = ast.tasks().collect();
+        assert_eq!(tasks.len(), 1);
+
+        let decls: Vec<_> = tasks[0].declarations().collect();
+        assert_eq!(decls.len(), 2);
+
+        let expr = decls[0].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 3);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  "),
+            _ => panic!("expected text part"),
+        }
+        match &stripped[1] {
+            StrippedStringPart::Placeholder(_) => {}
+            _ => panic!("expected interpolated part"),
+        }
+        match &stripped[2] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), ""),
+            _ => panic!("expected text part"),
+        }
+
+        let expr = decls[1].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 3);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "hello  "),
+            _ => panic!("expected text part"),
+        }
+        match &stripped[1] {
+            StrippedStringPart::Placeholder(_) => {}
+            _ => panic!("expected interpolated part"),
+        }
+        match &stripped[2] {
+            StrippedStringPart::Text(text) => assert_eq!(text.as_str(), "\nmy name\nis Jerry!"),
+            _ => panic!("expected text part"),
+        }
+    }
+
+    #[test]
+    fn remove_multiple_line_continuations() {
+        let (document, diagnostics) = Document::parse(
+            r#"
+version 1.2
+
+task test {
+    String hw = <<<
+    hello world \
+    \
+    \
+    my name is Jeff.
+    >>>
+}"#,
+        );
+
+        assert!(diagnostics.is_empty());
+        let ast = document.ast();
+        let ast = ast.as_v1().expect("should be a V1 AST");
+
+        let tasks: Vec<_> = ast.tasks().collect();
+        assert_eq!(tasks.len(), 1);
+
+        let decls: Vec<_> = tasks[0].declarations().collect();
+        assert_eq!(decls.len(), 1);
+
+        let expr = decls[0].expr().unwrap_literal().unwrap_string();
+        let stripped = expr.strip_whitespace().unwrap();
+        assert_eq!(stripped.len(), 1);
+        match &stripped[0] {
+            StrippedStringPart::Text(text) => {
+                assert_eq!(text.as_str(), "hello world my name is Jeff.")
+            }
+            _ => panic!("expected text part"),
+        }
     }
 }